# Conflicts:
#	server/src/main/java/org/apache/iotdb/db/mpp/plan/analyze/ColumnPaginationController.java

+++ metrics/interface/src/main/java/org/apache/iotdb/metrics/metricsets/disk/IDiskMetricsManager.java
@@ -24,11 +24,11 @@ import java.util.Map;
 import java.util.Set;
 
 public interface IDiskMetricsManager {
-  default Map<String, Long> getReadDataSizeForDisk() {
+  default Map<String, Double> getReadDataSizeForDisk() {
     return Collections.emptyMap();
   }
 
-  default Map<String, Long> getWriteDataSizeForDisk() {
+  default Map<String, Double> getWriteDataSizeForDisk() {
     return Collections.emptyMap();
   }
 
@@ -56,7 +56,7 @@ public interface IDiskMetricsManager {
     return Collections.emptyMap();
   }
 
-  default Map<String, Long> getIoUtilsPercentage() {
+  default Map<String, Double> getIoUtilsPercentage() {
     return Collections.emptyMap();
   }
 
@@ -76,16 +76,16 @@ public interface IDiskMetricsManager {
     return Collections.emptyMap();
   }
 
-  default Map<String, Long> getQueueSizeForDisk() {
+  default Map<String, Double> getQueueSizeForDisk() {
     return Collections.emptyMap();
   }
 
-  default long getActualReadDataSizeForProcess() {
-    return 0L;
+  default double getActualReadDataSizeForProcess() {
+    return 0.0;
   }
 
-  default long getActualWriteDataSizeForProcess() {
-    return 0L;
+  default double getActualWriteDataSizeForProcess() {
+    return 0.0;
   }
 
   default long getReadOpsCountForProcess() {
@@ -96,12 +96,12 @@ public interface IDiskMetricsManager {
     return 0L;
   }
 
-  default long getAttemptReadSizeForProcess() {
-    return 0L;
+  default double getAttemptReadSizeForProcess() {
+    return 0.0;
   }
 
-  default long getAttemptWriteSizeForProcess() {
-    return 0L;
+  default double getAttemptWriteSizeForProcess() {
+    return 0.0;
   }
 
   default Set<String> getDiskIds() {
