+++ server/src/main/java/org/apache/iotdb/db/metadata/plan/schemaregion/read/IShowSchemaPlan.java
@@ -39,9 +39,9 @@ public interface IShowSchemaPlan extends ISchemaRegionPlan {
 
   PartialPath getPath();
 
-  int getLimit();
+  long getLimit();
 
-  int getOffset();
+  long getOffset();
 
   boolean isPrefixMatch();
 }
